requirements.txt
PyPDF2>=3.0.0
Pillow>=10.0.0
flask>=2.0.0
werkzeug>=2.0.0 